+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### Spring Web/MVC
 
+- Added inspection for supported version of OpenApi
+- Added completion in openapi specification
 - Automatically apply OpenAPI JSON schema (Specification 3.0.0/3.1.0)
 - Supported dark theme for SwaggerUI panel
 - Added navigation from openapi http request to preview (SwaggerUI panel)